refactor: extract DataSelect as a component
