+++ admin/react/src/app.jsx
@@ -1,39 +1,36 @@
 import { useEffect, useState } from 'react'
-import Table from './components/Table'
 import MapSettings from './components/MapSettings'
 import { createId } from '@paralleldrive/cuid2'
 import MapList from './components/MapList'
 import { formDefaults } from './data/formDefaults'
-import {
-  deleteWpNodes,
-  getWpMaps,
-  saveCustomNodes,
-  saveWpMap,
-  saveWpNodes,
-  updateCustomNodes,
-  updateCustomNodesStatus,
-  updateWpMap,
-  updateWpNodes
-} from './utils/api'
+import { getWpMaps, saveCustomNodes, saveWpMap, updateWpMap } from './utils/api'
 import DataSource from './components/DataSource'
 import ProgressBar from './components/ProgressBar'
+import DataSelect from './components/DataSelect'
 
 export default function App() {
   // eslint-disable-next-line no-undef
   const wordpressUrl = murmurations_aggregator.wordpress_url
   const apiUrl = `${wordpressUrl}/wp-json/murmurations-aggregator/v1`
 
-  const [formData, setFormData] = useState(formDefaults)
-  const [maps, setMaps] = useState([])
-  const [selectedCountry, setSelectedCountry] = useState([])
-  const [profileList, setProfileList] = useState([])
-  const [selectedIds, setSelectedIds] = useState([])
+  // button states
   const [isLoading, setIsLoading] = useState(false)
-  const [progress, setProgress] = useState(0)
-  const [isEdit, setIsEdit] = useState(false)
-  const [selectedStatusOption, setSelectedStatusOption] = useState('publish')
   const [isRetrieving, setIsRetrieving] = useState(false)
+  const [isEdit, setIsEdit] = useState(false)
+
+  // MapList states
+  const [maps, setMaps] = useState([])
+
+  // DataSource states
+  const [formData, setFormData] = useState(formDefaults)
+  const [selectedCountry, setSelectedCountry] = useState([])
   const [tagSlug, setTagSlug] = useState(null)
+
+  // Table states
+  const [profileList, setProfileList] = useState([])
+
+  // ProgressBar states
+  const [progress, setProgress] = useState(0)
 
   useEffect(() => {
     getMaps().then(() => {
@@ -64,23 +61,6 @@ export default function App() {
       alert(
         `Error getting maps, please contact the administrator, error: ${error}`
       )
-    }
-  }
-
-  // Table
-  const toggleSelectAll = () => {
-    if (selectedIds.length === profileList.length) {
-      setSelectedIds([])
-    } else {
-      setSelectedIds(profileList.map(response => response.id))
-    }
-  }
-
-  const toggleSelect = id => {
-    if (selectedIds.includes(id)) {
-      setSelectedIds(selectedIds.filter(selectedId => selectedId !== id))
-    } else {
-      setSelectedIds([...selectedIds, id])
     }
   }
 
@@ -232,240 +212,6 @@ export default function App() {
     }
   }
 
-  const handleProfilesSubmit = async event => {
-    event.preventDefault()
-    setIsLoading(true)
-
-    try {
-      // get the selected profiles
-      const selectedProfiles = profileList.filter(profile =>
-        selectedIds.includes(profile.id)
-      )
-      const progressStep = 100 / selectedProfiles.length
-
-      for (let i = 0; i < selectedProfiles.length; i++) {
-        // update progress
-        if ((i + 1) * progressStep > 100) {
-          setProgress(100)
-        } else {
-          setProgress((i + 1) * progressStep)
-        }
-
-        const profile = selectedProfiles[i]
-
-        // if the profile wants to publish, it will create post in WordPress
-        // otherwise, dismiss, ignore status will update the status of nodes table
-        let profileResponse
-        if (selectedStatusOption === 'publish') {
-          profileResponse = await saveWpNodes(apiUrl, profile.tag_slug, profile)
-        } else {
-          profileResponse = await updateCustomNodesStatus(
-            apiUrl,
-            profile.map_id,
-            profile.profile_url,
-            selectedStatusOption
-          )
-        }
-
-        // todo: needs to summarize errors and display them in once
-        if (!profileResponse.ok) {
-          const profileResponseData = await profileResponse.json()
-          alert(
-            `Profile Error: ${profileResponse.status} ${JSON.stringify(
-              profileResponseData
-            )}`
-          )
-        }
-      }
-
-      // remove selected profiles
-      const newProfileList = profileList.filter(
-        profile => !selectedIds.includes(profile.id)
-      )
-
-      // if the extra_notes of all profiles are unavailable, it means all nodes are handled, we can refresh the page
-      if (
-        newProfileList.length === 0 ||
-        newProfileList.every(profile => profile.extra_notes === 'unavailable')
-      ) {
-        setFormData(formDefaults)
-        setSelectedCountry([])
-        setProfileList([])
-        await getMaps()
-      } else {
-        setProfileList(newProfileList)
-      }
-    } catch (error) {
-      alert(
-        `Handle Profiles Submit error: ${JSON.stringify(
-          error
-        )}, please delete the map and retry again.`
-      )
-    } finally {
-      // set everything back to default
-      setSelectedIds([])
-      setProgress(0)
-      setIsLoading(false)
-    }
-  }
-
-  const handleRetrieveProfilesSubmit = async event => {
-    event.preventDefault()
-    setIsLoading(true)
-
-    try {
-      // get the selected profiles
-      const selectedProfiles = profileList.filter(profile =>
-        selectedIds.includes(profile.id)
-      )
-      const progressStep = 100 / selectedProfiles.length
-
-      for (let i = 0; i < selectedProfiles.length; i++) {
-        // update progress
-        if ((i + 1) * progressStep > 100) {
-          setProgress(100)
-        } else {
-          setProgress((i + 1) * progressStep)
-        }
-
-        const profile = selectedProfiles[i]
-
-        // need to update the node first
-        if (profile.extra_notes === 'see updates') {
-          const profileResponse = await updateCustomNodes(
-            apiUrl,
-            profile.map_id,
-            profile.profile_url,
-            profile.profile_data
-          )
-
-          if (!profileResponse.ok) {
-            const profileResponseData = await profileResponse.json()
-            alert(
-              `Update Profile Error: ${profileResponse.status} ${JSON.stringify(
-                profileResponseData
-              )}`
-            )
-          }
-        }
-
-        const profileStatus = profile.status
-        // if original status and selected status are both publish, we need to update the post
-        if (selectedStatusOption === 'publish' && profileStatus === 'publish') {
-          const profileResponse = await updateWpNodes(apiUrl, profile)
-
-          if (!profileResponse.ok) {
-            const profileResponseData = await profileResponse.json()
-            alert(
-              `Update Profile Error: ${profileResponse.status} ${JSON.stringify(
-                profileResponseData
-              )}`
-            )
-          }
-        }
-
-        if (
-          selectedStatusOption === 'publish' &&
-          (profileStatus === 'new' || profileStatus === 'dismiss')
-        ) {
-          const profileResponse = await saveWpNodes(
-            apiUrl,
-            profile.tag_slug,
-            profile
-          )
-          if (!profileResponse.ok) {
-            const profileResponseData = await profileResponse.json()
-            alert(
-              `Create Profile Error: ${profileResponse.status} ${JSON.stringify(
-                profileResponseData
-              )}`
-            )
-          }
-        }
-
-        if (
-          (selectedStatusOption === 'dismiss' ||
-            selectedStatusOption === 'ignore') &&
-          profileStatus === 'publish'
-        ) {
-          const response = await deleteWpNodes(apiUrl, profile)
-
-          if (!response.ok) {
-            const responseData = await response.json()
-            alert(
-              `Delete Profile Error: ${response.status} ${JSON.stringify(
-                responseData
-              )}`
-            )
-          }
-
-          // update the status of the node
-          const profileResponse = await updateCustomNodesStatus(
-            apiUrl,
-            profile.map_id,
-            profile.profile_url,
-            selectedStatusOption
-          )
-          if (!profileResponse.ok) {
-            const profileResponseData = await profileResponse.json()
-            alert(
-              `Node Error: ${profileResponse.status} ${JSON.stringify(
-                profileResponseData
-              )}`
-            )
-          }
-        }
-
-        if (selectedStatusOption === 'ignore' && profileStatus === 'dismiss') {
-          // update the status of the node
-          const profileResponse = await updateCustomNodesStatus(
-            apiUrl,
-            profile.map_id,
-            profile.profile_url,
-            selectedStatusOption
-          )
-          if (!profileResponse.ok) {
-            const profileResponseData = await profileResponse.json()
-            alert(
-              `Node Error: ${profileResponse.status} ${JSON.stringify(
-                profileResponseData
-              )}`
-            )
-          }
-        }
-      }
-
-      // remove selected profiles
-      const newProfileList = profileList.filter(
-        profile => !selectedIds.includes(profile.id)
-      )
-
-      // if the extra_notes of all profiles are unavailable, it means all nodes are handled, we can refresh the page
-      if (
-        newProfileList.length === 0 ||
-        newProfileList.every(profile => profile.extra_notes === 'unavailable')
-      ) {
-        setFormData(formDefaults)
-        setSelectedCountry([])
-        setProfileList([])
-        await getMaps()
-      } else {
-        setProfileList(newProfileList)
-      }
-    } catch (error) {
-      alert(
-        `Handle Profiles Submit error: ${JSON.stringify(
-          error
-        )}, please delete the map and retry again.`
-      )
-    } finally {
-      // set everything back to default
-      setSelectedIds([])
-      setProgress(0)
-      setIsLoading(false)
-    }
-  }
-
   const handleEditSubmit = async event => {
     event.preventDefault()
     setIsLoading(true)
@@ -493,11 +239,6 @@ export default function App() {
       setTagSlug(null)
       await getMaps()
     }
-  }
-
-  const handleDropdownChange = function () {
-    const selected = document.getElementById('status-option').value
-    setSelectedStatusOption(selected)
   }
 
   return (
@@ -555,48 +296,19 @@ export default function App() {
               </div>
             )
           ) : (
-            <div>
-              <h2 className="text-xl mt-4">Data Select</h2>
-              <Table
-                tableList={profileList}
-                selectedIds={selectedIds}
-                onSelectAll={toggleSelectAll}
-                onSelect={toggleSelect}
-              />
-              <form
-                onSubmit={
-                  isRetrieving
-                    ? handleRetrieveProfilesSubmit
-                    : handleProfilesSubmit
-                }
-                className="p-6"
-              >
-                <div className="mt-6">
-                  {isLoading && <ProgressBar progress={progress} />}
-                  <div className="flex items-start">
-                    <label className="mr-2">Select Action:</label>
-                    <select
-                      id="status-option"
-                      value={selectedStatusOption}
-                      onChange={() => handleDropdownChange()}
-                      className="mr-2"
-                    >
-                      <option value="publish">Publish</option>
-                      <option value="dismiss">Dismiss</option>
-                      <option value="ignore">Ignore</option>
-                    </select>
-                    <button
-                      type="submit"
-                      className={`rounded-full bg-orange-500 px-4 py-2 font-bold text-white text-lg active:scale-90 hover:scale-110 hover:bg-orange-400 disabled:opacity-75 ${
-                        isLoading ? 'opacity-50 cursor-not-allowed' : ''
-                      }`}
-                    >
-                      {isLoading ? 'Submitting...' : 'Submit'}
-                    </button>
-                  </div>
-                </div>
-              </form>
-            </div>
+            <DataSelect
+              apiUrl={apiUrl}
+              profileList={profileList}
+              setProfileList={setProfileList}
+              isLoading={isLoading}
+              setIsLoading={setIsLoading}
+              isRetrieving={isRetrieving}
+              setProgress={setProgress}
+              progress={progress}
+              setFormData={setFormData}
+              setSelectedCountry={setSelectedCountry}
+              getMaps={getMaps}
+            />
           )}
         </div>
         <div className="w-1/2 mt-4 p-4">
