+++ admin/react/src/utils/filterProfile.js
@@ -1,0 +1,7 @@
+export const removeSelectedProfiles = (profileList, selectedIds) => {
+  return profileList.filter(profile => !selectedIds.includes(profile.id))
+}
+
+export const removeUnselectedProfiles = (profileList, selectedIds) => {
+  return profileList.filter(profile => selectedIds.includes(profile.id))
+}
